backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain-groq
